add a new slave function to set the slaveaddr based on hostname
A design with nginx is that it implements its own dns resolver.  This
resolver does not use the settings from resolv.conf, and so it does not
append a search path to the domain name.  Instead, we must resolve the
name another way, and we resolve this using the mesos state information.